requirements.txt
alembic==1.9.2
aniso8601==9.0.1
attrs==22.2.0
bcrypt==4.0.1
blinker==1.5
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-Migrate==4.0.2
flask-restx==1.0.5
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
SQLAlchemy==1.4.46
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
flask_session
flask_cors
azure-mgmt-storage
azure-storage-blob
psycopg2-binary